+++ editor/editor-plugin-annotation/src/pm-plugins/toolbar.tsx
@@ -227,6 +227,9 @@ export const buildToolbar: (editorAnalyticsAPI: EditorAnalyticsAPI | undefined) 
 									attributes: {
 										pageClass: 'editor',
 										commentType: 'inline',
+										...(fg('cc_comments_create_inline_experience_entry_point')
+											? { entryPoint: 'highlightActions' }
+											: {}),
 									},
 								});
 								createCommentExperience?.initExperience.start();
@@ -264,6 +267,9 @@ export const buildToolbar: (editorAnalyticsAPI: EditorAnalyticsAPI | undefined) 
 						attributes: {
 							pageClass: 'editor',
 							commentType: 'inline',
+							...(fg('cc_comments_create_inline_experience_entry_point')
+								? { entryPoint: 'highlightActions' }
+								: {}),
 						},
 					});
 					createCommentExperience?.initExperience.start();