+++ editor/renderer/src/ui/Renderer/index.tsx
@@ -773,7 +773,12 @@ const RendererWrapper = React.memo((props: RendererWrapperProps) => {
 					 * Telepointer changes are always a singular node-adds or node-removes.
 					 */
 					const isAdfStreamingEnabled =
-						expValEqualsNoExposure('platform_editor_ai_iw_adf_streaming', 'isEnabled', true) ||
+						expValEqualsNoExposure('platform_editor_ai_iw_adf_streaming', 'isEnabled', 'adf') ||
+						expValEqualsNoExposure(
+							'platform_editor_ai_iw_adf_streaming',
+							'isEnabled',
+							'adf_personalisation',
+						) ||
 						expValEqualsNoExposure('platform_editor_ai_ct_sg_adf_streaming', 'isEnabled', true);
 					if (
 						isAdfStreamingEnabled &&