+++ media/media-client/examples/1-uploader.tsx
@@ -67,6 +67,7 @@ class UploaderExample extends Component<UploaderExampleProps, UploaderExampleSta
 						<button onClick={this.onUploadStringClick}>Upload</button>
 					</div>
 					<div>
+						{/* eslint-disable-next-line @atlassian/a11y/aria-progressbar-name */}
 						<progress value={uploadingProgress} max="1" />
 					</div>
 					<div>Processing status: {processingStatus}</div>