+++ editor/editor-plugins/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @atlaskit/editor-plugins
 
+## 10.1.59
+
+### Patch Changes
+
+- Updated dependencies
+
+## 10.1.58
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 10.1.57
 
 ### Patch Changes